Move Parsers to SubFolder in Test
