+++ tests/Kinepolis/Parser/KinepolisPriceParserTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Kinepolis;
+namespace CultuurNet\UDB3\Kinepolis\Parser;
 
-use CultuurNet\UDB3\Kinepolis\Parser\KinepolisPriceParser;
+use CultuurNet\UDB3\Kinepolis\ParsedPriceForATheater;
 use PHPUnit\Framework\TestCase;
 
 final class KinepolisPriceParserTest extends TestCase
